+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0-beta] - 2022-08-17
+
+### Added
+
+- Expansion panels for groups in Known
+- Text search includes group names
+
 ## [1.0.1-beta] - 2022-08-16
 
 ### Fixed